### 2016-10-27 #101 @whedon
[PRE REVIEW]: ReproZip
**Submitting author:** @remram44 (<a href="http://orcid.org/0000-0002-0524-2282">Rémi Rampin</a>)
**Repository:** <a href="https://github.com/ViDA-NYU/reprozip" target ="_blank">https://github.com/ViDA-NYU/reprozip</a>
**Version:** 1.0.8
