### 2016-10-27 #101 @arfon
👋  @acabunoc - would you be able to edit this for us? It looks like `@jsta` might be a good reviewer for this one: https://github.com/openjournals/joss/blob/master/docs/reviewers.csv#L22
