### 2016-10-27 #101 @remram44
Review is for [version 1.0.8](https://github.com/ViDA-NYU/reprozip/tree/1.0.8), [10.5281/zenodo.159604](https://doi.org/10.5281/zenodo.159604). Required paper.md is [in branch `joss-paper`](https://github.com/ViDA-NYU/reprozip/tree/joss-paper/paper).
