### 2016-11-03 #101 @jsta
Go head @stain.  This looks very interesting but I don't have time currently. I would be interested in hearing whether this tool might be easier for novices to use than Docker.
