### 2016-11-03 #107 @whedon
[REVIEW]: ReproZip
**Submitting author:** @remram44 (<a href="http://orcid.org/0000-0002-0524-2282">Rémi Rampin</a>)
**Repository:** <a href="https://github.com/ViDA-NYU/reprozip" target ="_blank">https://github.com/ViDA-NYU/reprozip</a>
**Version:** 1.0.8
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.159604" target="_blank">https://dx.doi.org/10.5281/zenodo.159604</a>
