### 2016-11-07 #107 @stain
Thanks for the submission,  @remram44 - [paper](https://github.com/ViDA-NYU/reprozip/blob/joss-paper/paper/paper.md) looks good!

I just need to complete the testing of the software - when I tried this weekend to rerun a reprozip with `chroot` I lost my terminal files in my actual `/dev` (!) - so need to try again and raise issue upstream if appropriate.

Using `pip install` was enough to get started, but I also want to just verify running from the [archived source](https://doi.org/10.5281/zenodo.159604).
