### 2016-11-07 #107 @remram44
Removing the folder you unpacked in will indeed remove files from `/dev`, since they are bind-mounted in there. We provide [a `destroy` command](http://reprozip.readthedocs.io/en/1.0.x/unpacking.html#removing-the-experiment-directory) that will deal with it, unmounting in the case of chroot, or destroying containers/VMs.

Thank you for your time and the quick review!
