### 2016-11-10 35ede16
README: Add Contribution section and License info
Suggested for openjournals/joss-reviews#107

You may want to extend the (c) Copyright to include all the [contributors](https://github.com/ViDA-NYU/reprozip/graphs/contributors) (and their institutions)
