### 2016-11-10 #107 @stain
*Community guidelines*:

The documentation already have a [developer guide](https://reprozip.readthedocs.io/en/1.0.x/developerguide.html) and - deep enough down - links to mailing lists, but it is not made clear in the github repo itself how to contribute, so I've suggested to add that in ViDA-NYU/reprozip#225

I also propose to add a **License** section to the README file as well, as not everyone recognize the [BSD 3 Clause](https://opensource.org/licenses/BSD-3-Clause) when they see it just in the [LICENSE.txt](https://github.com/ViDA-NYU/reprozip/blob/master/LICENSE.txt)
