### 2016-11-10 ViDA-NYU/reprozip#226 @stain
Improve References section in paper.md
Hi,

This is part of openjournals/joss-reviews#107

The [references here](https://github.com/ViDA-NYU/reprozip/blob/joss-paper/paper/paper.md#references) don't use DOIs where they should.

* `Archive on Zenodo` link should show and link to DOI  [doi:10.5281/zenodo.159604](https://doi.org/10.5281/zenodo.159604)
* https://github.com/ViDA-NYU/reprozip-examples should similarly be archived in Zenodo and given a DOI
* Is it possible to archive https://reprozip.readthedocs.io/en/1.0.x/ ? I believe it may be part of the `docs` folder of the existing Zenodo archive, but I'm not sure. I've added it to [web.archive.org](https://web.archive.org/web/20161110101921/https://reprozip.readthedocs.io/en/1.0.x/) - could you click-through to check all pages are there?
* "Most recent paper" should link to its [doi:10.1145/2882903.2899401](https://doi.org/10.1145/2882903.2899401)

Could you include authors and title for each of the references (in particular the paper)?  You don't need to go all classical journal style, just include the main authors (e.g. as shown in Zenodo). BTW, I like full name :-)
