### 2016-11-10 #107 @stain
@remram44 Just the two above issues in community guidelines ViDA-NYU/reprozip#225 and to improve the references ViDA-NYU/reprozip#226 - otherwise I think it's good to go!
