### 2016-11-10 ViDA-NYU/reprozip#226 @remram44
> https://github.com/ViDA-NYU/reprozip-examples should similarly be archived in Zenodo and given a DOI

We could register the OSF project, which has the GitHub project + the linked RPZ files (and the Vagrant machine).

> Is it possible to archive https://reprozip.readthedocs.io/en/1.0.x/ ? I believe it may be part of the docs folder of the existing Zenodo archive, but I'm not sure. I've added it to web.archive.org - could you click-through to check all pages are there?

The docs folder is indeed what the docs are built from, so that's archived.
