### 2016-11-10 ViDA-NYU/reprozip#226 @VickySteeves
Hi @stain !! Thanks for the suggestions. I'll push the changes ASAP.

> Archive on Zenodo link should show and link to DOI doi:10.5281/zenodo.159604

Done!

> https://github.com/ViDA-NYU/reprozip-examples should similarly be archived in Zenodo and given a DOI

We actually have a DOI of this via the [Open Science Framework](https://osf.io/8uxpv/registrations/). I will add this as soon as our last project administrator approves the registration.

> Is it possible to archive https://reprozip.readthedocs.io/en/1.0.x/ ? I believe it may be part of the docs folder of the existing Zenodo archive, but I'm not sure. I've added it to web.archive.org - could you click-through to check all pages are there?

The web.archive.org link doesn't actually have the pages, but as @remram44 said, it's archived with Zenodo.

>"Most recent paper" should link to its doi:10.1145/2882903.2899401

Done!

> Could you include authors and title for each of the references (in particular the paper)? You don't need to go all classical journal style, just include the main authors (e.g. as shown in Zenodo). BTW, I like full name :-)

Done!
