### 2016-11-10 e87965a
updated references section with DOIs and author names
