### 2016-11-10 ViDA-NYU/reprozip#226 @VickySteeves
Fixed [here](https://github.com/ViDA-NYU/reprozip/commit/e87965ac9d82601100d0f77d75e4bfabedc0c396).
