### 2016-11-14 ViDA-NYU/reprozip#226 @stain
Thanks! You can close this issue.
