### 2016-11-14 ViDA-NYU/reprozip#226 @stain
Perhaps including the Title as well as the author would be ideal.. but I'm satisfied with this.
