### 2016-11-14 #107 @stain
Sorry, @remram44 not sure if I should tick the `accepted` label, I think @acabunoc as the editor will do that formally.
