### 2016-11-15 #107 @acabunoc
This looks great! Happy to accept this.

Thanks so much @stain & @remram44 🎉 🎈 🍰 🎁

Flagging @arfon because I'm actually not sure what the last step is. In the meantime, could you make sure there's an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive?
