### 2016-11-16 #107 @acabunoc
Thanks for adding the link! Helpful.

> On Nov 15, 2016, at 5:57 PM, Remi Rampin <notifications@github.com> wrote:
>
> We have a DOI through Zenodo: 10.5281/zenodo.159604.
>
