### 2016-11-16 #107 @arfon
@remram44 - I don't seem to be able to find the `paper.md` (and associated `paper.bib`) file anywhere in this repository? Do you need to merge the `joss-paper` branch?
