### 2016-11-18 #107 @arfon
> The paper in on the joss-paper branch, does it need to be on master? It can't be added to the DOI'd archive retroactively anyway.

I think it should _eventually_ be on master (if that's your main branch).
