### 2016-11-29 #107 @remram44
@arfon: Is further action required on our part? I see that we haven't been added to joss-papers yet.
