### 2016-11-29 #107 @arfon
> @arfon: Is further action required on our part? I see that we haven't been added to joss-papers yet.

Sorry, this slipped through. On reflection I would like to see the `joss-paper` branch merged before I finalize this submission. Does that sound OK @remram44?
