### 2016-11-30 #107 @remram44
I remember talking about this with @jakevdp but he didn't really give an answer. I think I like it better this way, in its own branch directly based off the reviewed version (1.0.8). There is no need to clutter the repository (CI scripts are already bad enough), and it will not be updated.

Your joss-papers repository is the authoritative one anyway, and I'll add a badge linking to it in my master branch's README. Is that acceptable?
