### 2016-12-01 6b2f47a
Update paper.md
