### 2016-12-01 #107 @arfon
OK no problem @remram44. As long as the `joss-papers` branch is permanent this is OK with me.

I've opened a PR on your repo (https://github.com/ViDA-NYU/reprozip/pull/227) with some changes that need to be made to the paper structure before we can finish processing this submission. Please let me know when you've made these changes 😁
