### 2016-12-01 aec12ab
added paper.bib file
