+++ paper/paper.bib
@@ -1,0 +1,57 @@
+@online{ReproZip,
+  author = {R{\'e}mi  Rampin, Fernando Chrigiati, Vicky Steeves, Dennis Shasha, Juliana Freire},
+  title = {ReproZip: The Reproducibility Packer},
+  year = 2014,
+  url = {https://github.com/ViDA-NYU/reprozip},
+  doi = {https://doi.org/10.5281/zenodo.159604},
+  urldate = {2016-12-01},
+}
+
+@online{main-site,
+  author = {Fernando Chirigati, R{\'e}mi  Rampin, Vicky Steeves},
+  title = {ReproZip Website},
+  year = 2016,
+  url = {https://vida-nyu.github.io/reprozip/},
+  urldate = {2016-12-01},
+}
+
+@online{example-site,
+  author = {Vicky Steeves, R{\'e}mi  Rampin, Fernando Chirigati},
+  title = {ReproZip Examples Repository},
+  year = 2016,
+  url = {https://vida-nyu.github.io/reprozip-examples/},
+  doi = {http://doi.org/10.17605/OSF.IO/JB2UV},
+  urldate = {2016-12-01},
+}
+
+@misc{documentation,
+  author = {R{\'e}mi Rampin, Fernando Chrigiati, Vicky Steeves},
+  title = {ReproZip Documentation},
+  howpublished = {\url{https://reprozip.readthedocs.io/en/1.0.x/}},
+}
+
+@online{youtube,
+  title = {Packing and Unpacking Experiments with ReproZip},
+  year = 2015,
+  organization = {Youtube},
+  author = {Fernando Chirigati, Vicky Steeves},
+  url = {https://www.youtube.com/watch?v=-zLPuwCHXo0},
+ }
+
+@inproceedings{recent-paper,
+ author = {Chirigati, Fernando and Rampin, R{\'e}mi and Shasha, Dennis and Freire, Juliana},
+ title = {ReproZip: Computational Reproducibility With Ease},
+ booktitle = {Proceedings of the 2016 International Conference on Management of Data},
+ series = {SIGMOD '16},
+ year = {2016},
+ isbn = {978-1-4503-3531-7},
+ location = {San Francisco, California, USA},
+ pages = {2085--2088},
+ numpages = {4},
+ url = {http://doi.acm.org/10.1145/2882903.2899401},
+ doi = {10.1145/2882903.2899401},
+ acmid = {2899401},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {computational reproducibility, provenance, reprozip},
+} 
