### 2016-12-01 #107 @remram44
I merged [your changes](https://github.com/ViDA-NYU/reprozip/pull/227) and @VickySteeves added the requested [paper.bib](https://github.com/ViDA-NYU/reprozip/blob/joss-paper/paper/paper.bib).
