### 2016-12-02 #107 @arfon
@stain many thanks for the review here and thanks for editing this @acabunoc.

@remram44 - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00107 🚀 🎉 💥
